+++ src/Universalis.Application/Views/HistoryView.cs
@@ -21,7 +21,7 @@ namespace Universalis.Application.Views
         public string DcName { get; set; }
 
         [JsonProperty("lastUploadTime")]
-        public uint LastUploadTimeUnixMilliseconds { get; set; }
+        public double LastUploadTimeUnixMilliseconds { get; set; }
 
         [JsonProperty("stackSizeHistogram")]
         public SortedDictionary<int, int> StackSizeHistogram { get; set; } = new();
